requirements.txt
# Core dependencies
pytrends==4.9.2
pandas==2.1.4
numpy==1.24.3

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.15.0
prophet==1.1.5

# Data visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Utilities
python-dateutil==2.8.2
tqdm==4.66.1
joblib==1.3.2

# Configuration and logging
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development tools
black==23.12.0
flake8==6.1.0
isort==5.13.2 